import: factor out download.rs
