+++ src/data/import.rs
@@ -1,21 +1,14 @@
-use std::fs::File;
-use std::process::Command;
-
-use anyhow::{Context, Error};
+use anyhow::Error;
 use camino::{Utf8Path, Utf8PathBuf};
-use rust_embed::RustEmbed;
 use serde::{Deserialize, Serialize};
-use url::Url;
-
-use crate::download::{Download, Git, Source};
-use crate::utils::calculate_hash;
 
 mod cmd;
+mod download;
 
 #[derive(Debug, Serialize, Deserialize, Hash)]
 #[serde(untagged)]
 pub enum ImportEntry {
-    Download(Download),
+    Download(crate::download::Download),
     Command(cmd::Command),
 }
 
@@ -28,14 +21,12 @@ impl ImportEntry {
     }
 }
 
-#[derive(RustEmbed)]
-#[folder = "assets/imports"]
-struct Asset;
-
 pub trait Import: std::hash::Hash {
     fn get_name(&self) -> Option<String>;
     fn handle<T: AsRef<Utf8Path>>(&self, build_dir: T) -> Result<Utf8PathBuf, Error>;
     fn get_path<T: AsRef<Utf8Path>>(&self, build_dir: T) -> Result<Utf8PathBuf, Error> {
+        use crate::utils::calculate_hash;
+
         let source_hash = calculate_hash(&self);
 
         let mut res = Utf8PathBuf::from(build_dir.as_ref());
@@ -46,88 +37,6 @@ pub trait Import: std::hash::Hash {
             res.push(format!("{source_hash}"));
         }
         Ok(res)
-    }
-}
-
-impl Import for Download {
-    fn handle<T: AsRef<Utf8Path>>(&self, build_dir: T) -> Result<Utf8PathBuf, Error> {
-        let path = self.get_path(build_dir).unwrap();
-        let tagfile = path.join(".laze-downloaded");
-
-        if !tagfile.exists() {
-            match &self.source {
-                Source::Git(Git::Commit { url, commit }) => {
-                    println!("IMPORT Git {url}:{commit} -> {path}");
-
-                    let status = if Url::parse(url).is_ok() {
-                        Command::new("git")
-                            .args(["cache", "clone", url, commit, path.as_str()])
-                            .status()?
-                    } else {
-                        let mut status = Command::new("git")
-                            .args(["clone", "--no-checkout", url, path.as_str()])
-                            .status()?;
-                        if status.success() {
-                            status = Command::new("git")
-                                .args(["-C", path.as_str(), "checkout", commit])
-                                .status()?;
-                        }
-                        status
-                    };
-
-                    if status.success() {
-                        File::create(tagfile)?;
-                    } else {
-                        return Err(anyhow!(
-                            "could not import from git url: {url} commit: {commit}",
-                        ));
-                    }
-                }
-                Source::Laze(name) => {
-                    let mut at_least_one = false;
-                    let prefix = format!("{name}/");
-                    for filename in Asset::iter().filter(|x| x.starts_with(&prefix)) {
-                        if !at_least_one {
-                            at_least_one = true;
-                            std::fs::create_dir_all(&path)
-                                .with_context(|| format!("creating {path}"))?;
-                        }
-
-                        let embedded_file = Asset::get(&filename).unwrap();
-                        let filename = filename.strip_prefix(&prefix).unwrap();
-                        let filename = path.join(filename);
-                        let parent = path.parent().unwrap();
-                        std::fs::create_dir_all(path.parent().unwrap())
-                            .with_context(|| format!("creating {parent}"))?;
-                        std::fs::write(&filename, embedded_file.data)
-                            .with_context(|| format!("creating {filename}"))?;
-                    }
-                    if at_least_one {
-                        File::create(&tagfile).with_context(|| format!("creating {tagfile}"))?;
-                    } else {
-                        return Err(anyhow!("could not import from laze defaults: {name}"));
-                    }
-                }
-            }
-        }
-
-        get_lazefile(&path)
-    }
-
-    fn get_name(&self) -> Option<String> {
-        match &self.source {
-            Source::Git(Git::Commit { url, .. }) => url.split('/').last().map(|x| x.to_string()),
-            Source::Laze(name) => {
-                let prefix = format!("{}/", name);
-
-                if Asset::iter().any(|x| x.starts_with(&prefix)) {
-                    let build_uuid_hash = calculate_hash(&build_uuid::get());
-                    Some(format!("laze/{name}-{build_uuid_hash}"))
-                } else {
-                    None
-                }
-            } //_ => None,
-        }
     }
 }
